+++ lib/rails_wayback.rb
@@ -8,6 +8,7 @@ require "rails_wayback/render_tracker"
 require "rails_wayback/toggle"
 require "rails_wayback/git"
 require "rails_wayback/ref_policy"
+require "rails_wayback/cache_inventory"
 require "rails_wayback/view_source"
 require "rails_wayback/doctor"